Move test for concat() to extra file
